Remove some compiler warnings.
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11172 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXmediaLib/src/mp/MpRawAudioBuffer.cpp
@@ -58,7 +58,7 @@ MpRawAudioBuffer::MpRawAudioBuffer(const char* pFilePath)
    pAudioAbstract->minMaxSamplingRate(&minRate, &maxRate, &prefRate);
    if (prefRate != 8000) {
       OsSysLog::add(FAC_MP, PRI_ERR, "MpRawAudioBuffer::MpRawAudioBuffer(%s) Invalid sampling rate: %d",
-                    pFilePath, prefRate);
+                    pFilePath, (int) prefRate);
       return;
    }
 
@@ -80,7 +80,7 @@ MpRawAudioBuffer::MpRawAudioBuffer(const char* pFilePath)
    mpAudioBuffer = new char[mAudioBufferLength];
    if (mpAudioBuffer == NULL) {
       OsSysLog::add(FAC_MP, PRI_ERR, "MpRawAudioBuffer::MpRawAudioBuffer(%s) - Failed to allocate storage: new char[%d]",
-                    pFilePath, mAudioBufferLength);
+                    pFilePath, (int) mAudioBufferLength);
       mAudioBufferLength = 0;
       return;
    }
